requirements.txt
Django==3.1.1
psycopg2==2.8.6
django-rest-framework-0.1.0